+++ src/orquestra/sdk/_base/cli/_dorq/_ui/_presenters.py
@@ -11,14 +11,14 @@ import sys
 import typing as t
 import webbrowser
 from contextlib import contextmanager
-from datetime import datetime, timezone
+from datetime import datetime
 from pathlib import Path
 from typing import Iterable, Iterator, List, Sequence
 
 import click
 from tabulate import tabulate
 
-from orquestra.sdk._base import _config, _env, _services, serde
+from orquestra.sdk._base import _config, _dates, _env, _services, serde
 from orquestra.sdk.schema import responses
 from orquestra.sdk.schema.ir import ArtifactFormat
 from orquestra.sdk.schema.workflow_run import (
@@ -294,7 +294,7 @@ class PromptPresenter:
             wfs,
             key=lambda wf: wf.status.start_time
             if wf.status.start_time
-            else datetime.fromtimestamp(0).replace(tzinfo=timezone.utc),
+            else _dates.from_unix_time(0),
             reverse=True,
         )