+++ src/orquestra/sdk/_ray/_dag.py
@@ -21,7 +21,7 @@ import pydantic
 from orquestra.sdk.schema.responses import WorkflowResult
 
 from .. import exceptions
-from .._base import _services, serde
+from .._base import _dates, _services, serde
 from .._base._db import WorkflowDB
 from .._base._env import RAY_GLOBAL_WF_RUN_ID_ENV
 from .._base._logs._interfaces import LogReader
@@ -48,13 +48,9 @@ from ._wf_metadata import InvUserMetadata, WfUserMetadata, pydatic_to_json_dict
 
 
 def _instant_from_timestamp(unix_timestamp: t.Optional[float]) -> t.Optional[datetime]:
-    """
-    Parses a unix epoch timestamp (UTC seconds since 1970) into a
-    timezone-aware datetime object.
-    """
     if unix_timestamp is None:
         return None
-    return datetime.fromtimestamp(unix_timestamp, timezone.utc)
+    return _dates.from_unix_time(unix_timestamp)
 
 
 def _generate_wf_run_id(wf_def: ir.WorkflowDef):
@@ -514,7 +510,7 @@ class RayRuntime(RuntimeInterface):
             raise exceptions.WorkspacesNotSupportedError(
                 "Filtering by workspace or project is not supported on Ray runtimes."
             )
-        now = datetime.now(timezone.utc)
+        now = _dates.now()
 
         if state is not None:
             if not isinstance(state, list):